requirements.txt
requests>=2.28.0
google-generativeai>=0.2.0
fastapi>=0.95.0
uvicorn>=0.22.0
